--- __unported__/account_tax_expense_include/account_tax_expense_include.py
@@ -1,89 +1,0 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    Odoo, Open Source Management Solution
-#    Copyright (C) 2010 - 2014 Savoir-faire Linux
-#    (<http://www.savoirfairelinux.com>).
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp.osv import orm, fields
-from openerp.tools.float_utils import float_round
-
-
-class account_tax(orm.Model):
-    _inherit = 'account.tax'
-    _columns = {
-        'expense_include': fields.boolean('Tax Included in Expense',
-                                          help="Check this if this tax is \
-                                            included in the expense amount."),
-    }
-
-    def compute_all(self, cr, uid, taxes, price_unit, quantity, product=None, partner=None, force_excluded=False):
-        """
-        :param force_excluded: boolean used to say that we don't want to consider the value of field price_include of
-            tax. It's used in encoding by line where you don't matter if you encoded a tax with that boolean to True or
-            False
-        RETURN: {
-                'total': 0.0,                # Total without taxes
-                'total_included: 0.0,        # Total with taxes
-                'taxes': []                  # List of taxes, see compute for the format
-            }
-        """
-
-        # By default, for each tax, tax amount will first be computed
-        # and rounded at the 'Account' decimal precision for each
-        # PO/SO/invoice line and then these rounded amounts will be
-        # summed, leading to the total amount for that tax. But, if the
-        # company has tax_calculation_rounding_method = round_globally,
-        # we still follow the same method, but we use a much larger
-        # precision when we round the tax amount for each line (we use
-        # the 'Account' decimal precision + 5), and that way it's like
-        # rounding after the sum of the tax amounts of each line
-        precision = self.pool.get('decimal.precision').precision_get(cr, uid, 'Account')
-        tax_compute_precision = precision
-        if taxes and taxes[0].company_id.tax_calculation_rounding_method == 'round_globally':
-            tax_compute_precision += 5
-        totalin = totalex = float_round(price_unit * quantity, precision)
-        tin = []
-        tex = []
-        for tax in taxes:
-            if not tax.price_include or force_excluded:
-                tex.append(tax)
-            else:
-                tin.append(tax)
-        tin = self.compute_inv(
-            cr, uid, tin, price_unit, quantity, product=product,
-            partner=partner, precision=tax_compute_precision
-        )
-        for r in tin:
-            totalex -= r.get('amount', 0.0)
-        totlex_qty = 0.0
-        try:
-            totlex_qty = totalex / quantity
-        except:
-            pass
-        tex = self._compute(
-            cr, uid, tex, totlex_qty, quantity, product=product,
-            partner=partner, precision=tax_compute_precision
-        )
-        for r in tex:
-            totalin += r.get('amount', 0.0)
-        return {
-            'total': totalex,
-            'total_included': totalin,
-            'taxes': tin + tex
-        }
