+++ account_tax_expense_include/account_tax_expense_include_view.xml
@@ -14,17 +14,6 @@
             </field>
         </record>
 
-       <record id="view_partner_form" model="ir.ui.view">
-            <field name="name">res.partner.form.expense_include</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="base.view_partner_form"/>
-            <field name="arch" type="xml">
-                 <field name="supplier" position="after" >
-                      <field name="employee"/>
-                 </field>
-            </field>
-        </record>
-
     </data>
 </openerp>
